requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
decorator==5.1.1
dnspython==2.4.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-search-results==2.4.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandas==2.0.3
pdfminer.six==20221105
Pillow==9.5.0
pinecone-client==2.2.2
protobuf==4.23.4
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.1
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
yarl==1.9.2
zipp==3.16.2
